Initial commit for rework of Omer
