+++ hdate/date.py
@@ -16,24 +16,18 @@ from hdate import htables
 from hdate.gematria import hebrew_number
 from hdate.hebrew_date import HebrewDate
 from hdate.htables import Days, Holiday, HolidayTypes, Months, Parasha
+from hdate.omer import Omer
 from hdate.translator import TranslatorMixin
 
 _LOGGER = logging.getLogger(__name__)
 
 
-class HDate(TranslatorMixin):
+class HDate(TranslatorMixin):  # pylint: disable=too-many-instance-attributes
     """
     Hebrew date class.
 
     Supports converting from Gregorian and Julian to Hebrew date.
     """
-
-    # Prefixes and strings for different languages
-    OMER_SUFFIX = {
-        "hebrew": "בעומר",
-        "english": "in the Omer",
-        "french": "jour du Omer",
-    }
 
     def __init__(
         self,
@@ -43,6 +37,7 @@ class HDate(TranslatorMixin):
         heb_date: Optional[HebrewDate] = None,
     ) -> None:
         """Initialize the HDate object."""
+        self.language = language
         super().__init__()
         # Initialize private variables
         self._last_updated = ""
@@ -55,7 +50,7 @@ class HDate(TranslatorMixin):
             self._gdate = heb_date.to_gdate()
 
         self.diaspora = diaspora
-        self.set_language(language)
+        self.omer = Omer(date=self.hdate, language=language)
 
     def __str__(self) -> str:
         """Return a full Unicode representation of HDate."""
@@ -64,11 +59,8 @@ class HDate(TranslatorMixin):
         year_number = hebrew_number(self.hdate.year, language=self._language)
         result = f"{self.dow} {day_number} {in_prefix}{self.hdate.month} {year_number}"
 
-        # Handle Omer day
-        if 0 < self.omer_day < 50:
-            omer_day_number = hebrew_number(self.omer_day, language=self._language)
-            omer_suffix = self.OMER_SUFFIX.get(self._language, "in the Omer")
-            result = f"{result} {omer_day_number} {omer_suffix}"
+        if self.omer.total_days > 0:
+            result = f"{result} {self.omer}"
 
         # Append holiday description if any
         if self.holidays:
@@ -115,6 +107,7 @@ class HDate(TranslatorMixin):
 
         self._last_updated = "hdate"
         self._hdate = date
+        self.omer = Omer(date=self.hdate, language=self._language)
 
     @property
     def gdate(self) -> dt.date:
@@ -185,15 +178,6 @@ class HDate(TranslatorMixin):
         dow = Days(_dow)
         dow.set_language(self._language)
         return dow
-
-    @property
-    def omer_day(self) -> int:
-        """Return the day of the Omer."""
-        first_omer_day = HebrewDate(self.hdate.year, Months.NISAN, 16)
-        omer_day = (self.hdate - first_omer_day).days + 1
-        if not 0 < omer_day < 50:
-            return 0
-        return omer_day
 
     @property
     def daf_yomi(self) -> str:
@@ -425,224 +409,4 @@ class HDate(TranslatorMixin):
             and self.hdate.year < self.upcoming_shabbat.hdate.year
         ):
             return self.upcoming_shabbat.get_reading()
-        return cast(Parasha, readings[weeks])
-
-
-def get_omer_string(omer: int, language: str = "hebrew") -> str:
-    """Return a string representing the count of the Omer."""
-
-    if not 0 < omer < 50:
-        raise ValueError(f"Invalid Omer day: {omer}")
-
-    if language == "hebrew":
-        return _get_omer_string_hebrew(omer)
-    if language == "english":
-        return _get_omer_string_english(omer)
-    if language == "french":
-        return _get_omer_string_french(omer)
-
-    return f"Today is day {omer} of the Omer."
-
-
-def _get_omer_string_hebrew(omer: int) -> str:
-    """Return a string representing the count of the Omer in hebrew."""
-    tens = ["", "עשרה", "עשרים", "שלושים", "ארבעים"]
-    ones = [
-        "",
-        "אחד",
-        "שנים",
-        "שלושה",
-        "ארבעה",
-        "חמשה",
-        "ששה",
-        "שבעה",
-        "שמונה",
-        "תשעה",
-    ]
-
-    omer_string = "היום "
-
-    if 10 < omer < 20:
-        omer_string += f"{ones[omer % 10]} עשר"
-    elif omer >= 10:
-        unit_part = ones[omer % 10]
-        if omer % 10:
-            unit_part += " ו"
-        omer_string += unit_part + tens[omer // 10]
-    else:
-
-        if omer > 2:
-            omer_string += ones[omer]
-
-    if omer > 2:
-        if omer < 11:
-            omer_string += " ימים "
-        else:
-            omer_string += " יום "
-    elif omer == 1:
-        omer_string += "יום אחד "
-    else:  # omer == 2
-        omer_string += "שני ימים "
-
-    if omer > 6:
-        omer_string += "שהם "
-        weeks, days = divmod(omer, 7)
-
-        week_mapping = {1: "שבוע אחד ", 2: "שני שבועות "}
-        omer_string += week_mapping.get(weeks, f"{ones[weeks]} שבועות ")
-
-        if days:
-            day_mapping = {1: "יום אחד ", 2: "שני ימים "}
-            omer_string += "ו" + day_mapping.get(days, f"{ones[days]} ימים ")
-
-    omer_string += "לעומר"
-    return omer_string
-
-
-def _get_omer_string_english(omer: int) -> str:
-    """Return a string representing the count of the Omer in english."""
-    ones = [
-        "",
-        "one",
-        "two",
-        "three",
-        "four",
-        "five",
-        "six",
-        "seven",
-        "eight",
-        "nine",
-    ]
-    teens = [
-        "ten",
-        "eleven",
-        "twelve",
-        "thirteen",
-        "fourteen",
-        "fifteen",
-        "sixteen",
-        "seventeen",
-        "eighteen",
-        "nineteen",
-    ]
-    tens = ["", "", "twenty", "thirty", "forty"]
-    omer_string = "Today is "
-    if omer < 10:
-        omer_string += ones[omer]
-    elif 10 <= omer < 20:
-        omer_string += teens[omer - 10]
-    else:
-        ten = omer // 10
-        one = omer % 10
-        omer_string += tens[ten]
-        if one != 0:
-            omer_string += "-" + ones[one]
-    omer_string += " day"
-    if omer != 1:
-        omer_string += "s"
-    omer_string += " of the Omer"
-    # Add weeks and days
-    weeks = omer // 7
-    days = omer % 7
-    if weeks > 0:
-        omer_string += f", which is {weeks} week"
-        if weeks != 1:
-            omer_string += "s"
-        if days > 0:
-            omer_string += f" and {days} day"
-            if days != 1:
-                omer_string += "s"
-    omer_string += "."
-    return omer_string
-
-
-def _get_omer_string_french(omer: int) -> str:
-    """Return a string representing the count of the Omer in french."""
-    ones = [
-        "",
-        "un",
-        "deux",
-        "trois",
-        "quatre",
-        "cinq",
-        "six",
-        "sept",
-        "huit",
-        "neuf",
-    ]
-    teens = [
-        "dix",
-        "onze",
-        "douze",
-        "treize",
-        "quatorze",
-        "quinze",
-        "seize",
-        "dix-sept",
-        "dix-huit",
-        "dix-neuf",
-    ]
-    tens = ["", "", "vingt", "trente", "quarante"]
-
-    irregular_ordinals = {
-        1: "premier",
-        2: "deuxième",
-        3: "troisième",
-        4: "quatrième",
-        5: "cinquième",
-        6: "sixième",
-        7: "septième",
-        8: "huitième",
-        9: "neuvième",
-        10: "dixième",
-        11: "onzième",
-        12: "douzième",
-        13: "treizième",
-        14: "quatorzième",
-        15: "quinzième",
-        16: "seizième",
-        20: "vingtième",
-        30: "trentième",
-        40: "quarantième",
-    }
-
-    # Init
-    omer_string = "Aujourd'hui c'est le "
-
-    ten = omer // 10
-    one = omer % 10
-
-    # Construction
-    if omer in irregular_ordinals:
-        ordinal = irregular_ordinals[omer]
-    else:
-        if omer < 10:
-            number_word = ones[omer]
-        elif 10 < omer < 20:
-            number_word = teens[omer - 10]
-        else:
-            if one == 1 and ten > 1:
-                number_word = tens[ten] + " et un"
-            else:
-                number_word = tens[ten]
-                if one != 0:
-                    number_word += "-" + ones[one]
-
-        ordinal = number_word + "ième"
-
-    omer_string += ordinal + " jour de l'Omer"
-    # Add weeks and days
-    weeks = omer // 7
-    days = omer % 7
-
-    if weeks > 0:
-        omer_string += f", ce qui fait {weeks} semaine"
-        if weeks != 1:
-            omer_string += "s"
-        if days > 0:
-            omer_string += f" et {days} jour"
-            if days != 1:
-                omer_string += "s"
-
-        omer_string += "."
-    return omer_string
+        return cast(htables.Parasha, readings[weeks])
