+++ hdate/translations/en.json
@@ -201,5 +201,8 @@
         "barech_aleinu": "Barech aleinu",
         "veten_tal": "VeTen Tal uMatar Livracha",
         "veten_beracha": "VeTen Beracha"
+    },
+    "Omer": {
+        "in_omer": "in the Omer"
     }        
 }
